shengji/requirements/global.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/global.txt requirements/global.in
#
appdirs==1.4.3            # via setuptools
asgiref==1.0.1            # via channels, daphne
attrs==16.3.0             # via automat
autobahn==0.18.0          # via daphne
automat==0.5.0            # via twisted
channels==1.1.1
constantly==15.1.0        # via twisted
daphne==1.1.0             # via channels
dj-database-url==0.4.2
django-admin-decorators==0.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.7
django-dotenv==1.4.1
django-extensions==1.7.7
django-logentry-admin==1.0.2
django-registration==2.2
django-reversion==2.0.8
django==1.10.6
incremental==16.10.1      # via twisted
packaging==16.8           # via setuptools
pyparsing==2.2.0          # via packaging
pytz==2016.10
six==1.10.0               # via asgiref, autobahn, automat, django-extensions, packaging, setuptools, txaio
sqlparse==0.2.3           # via django-debug-toolbar
twisted==17.1.0           # via daphne
txaio==2.6.1              # via autobahn
zope.interface==4.3.3     # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via zope.interface
